pyproject.toml
[project]
name = "agentic-python-code-analyser"
version = "0.1.0"
description = ""
authors = [
    {name = "Cyril-Aubry",email = "76761963+Cyril-Aubry@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai (>=0.108.0,<0.109.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "crewai-tools (>=0.38.1,<0.39.0)"
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
